Minimal change to sim1.m to increase robustness
Previous commit used the algorithm now in lin_solve_robust.m by default
for solving the linear system in the Newton update in sim1.m.
This commit switches to a much simpler algorithm which just does
mldivide, then if that generates any nans or infs, replaces these
elements by zero, and tests if that is a solution.

Conflicts:
	matlab/utilities/tests